+++ okio/src/main/java/okio/BufferedSource.java
@@ -91,7 +91,7 @@ public interface BufferedSource extends Source {
    * Removes up to {@code byteCount} bytes from this and copies them into {@code sink} at
    * {@code offset}. Returns the number of bytes read, or -1 if this source is exhausted.
    */
-  int read(byte[] sink, int offset, long byteCount) throws IOException;
+  int read(byte[] sink, int offset, int byteCount) throws IOException;
 
   /**
    * Removes exactly {@code byteCount} bytes from this and appends them to
